WIP for writing info in one file.
