+++ src/main/java/com/googleintern/wfm/ruleengine/action/BUILD.bazel
@@ -9,18 +9,6 @@ java_binary(
 )
 
 java_binary(
-    name = "CsvWriter",
-    srcs = ["CsvWriter.java"],
-    main_class = "src.main.java.com.googleintern.wfm.ruleengine.action.CsvWriter",
-    visibility = ["//visibility:public"],
-    deps = [
-        ":maven_dependency",
-        "//src/main/java/com/googleintern/wfm/ruleengine/model:FilterModel",
-        "//src/main/java/com/googleintern/wfm/ruleengine/model:RuleModel",
-    ],
-)
-
-java_binary(
     name = "DataProcessor",
     srcs = ["DataProcessor.java"],
     visibility = ["//visibility:public"],
